requirements.txt
django
djangorestframework
pyppeteer
django-filter
